+++ src/common/cuepoint/index.js
@@ -1,0 +1,9 @@
+// @flow
+import {CUE_POINTS_TEXT_TRACK, CUE_POINT_KEY} from './cuepoint-type';
+
+const cuepoint = {
+  CUE_POINTS_TEXT_TRACK,
+  CUE_POINT_KEY
+};
+
+export {cuepoint};