+++ src/common/thumbnail-manager.js
@@ -1,5 +1,5 @@
 // @flow
-import {Utils, ThumbnailInfo, MediaType} from '@playkit-js/playkit-js';
+import {Utils, ThumbnailInfo, MediaType, EventManager} from '@playkit-js/playkit-js';
 import evaluate from './utils/evaluate';
 
 const DefaultThumbnailConfig: Object = {
@@ -12,19 +12,33 @@ const THUMBNAIL_REGEX = /.*\/p\/\d+\/(?:[a-zA-Z]+\/\d+\/)*thumbnail\/entry_id\/\
 const THUMBNAIL_SERVICE_TEMPLATE: string = '{{thumbnailUrl}}/width/{{thumbsWidth}}/vid_slices/{{thumbsSlices}}/ks/{{ks}}';
 
 class ThumbnailManager {
-  _player: Player;
+  _player: KalturaPlayer;
   _thumbnailConfig: ?KPThumbnailConfig;
+  _eventManager: EventManager;
+  _thumbsHeight: number;
 
-  constructor(player: Player, uiConfig: KPUIOptionsObject, mediaConfig: KPMediaConfig) {
+  constructor(player: KalturaPlayer, uiConfig: KPUIOptionsObject, mediaConfig: KPMediaConfig) {
     this._player = player;
     this._thumbnailConfig = this._buildKalturaThumbnailConfig(uiConfig, mediaConfig);
+    this._eventManager = new EventManager();
+    if (this._isUsingKalturaThumbnail()) {
+      const img = new Image();
+      this._eventManager.listenOnce(img, 'load', () => {
+        this._thumbsHeight = img.naturalHeight;
+      });
+      img.src = this._thumbnailConfig?.thumbsSprite || '';
+    }
+  }
+
+  destroy() {
+    this._eventManager.destroy();
   }
 
   getThumbnail(time: number): ?ThumbnailInfo {
     if (this._isUsingKalturaThumbnail()) {
       return this._convertKalturaThumbnailToThumbnailInfo(time);
     }
-    return this._player.getThumbnail(time);
+    return this._player._localPlayer.getThumbnail(time);
   }
 
   getKalturaThumbnailConfig(): ?KPThumbnailConfig {
@@ -38,13 +52,12 @@ class ThumbnailManager {
   _convertKalturaThumbnailToThumbnailInfo = (time: number): ?ThumbnailInfo => {
     if (this._thumbnailConfig) {
       const {thumbsSprite, thumbsWidth, thumbsSlices} = this._thumbnailConfig;
-      const {thumbsHeight} = DefaultThumbnailConfig;
       const duration = this._player.duration / thumbsSlices;
       const thumbnailInfo = {
         x: Math.floor(time / duration) * thumbsWidth,
-        y: thumbsHeight,
+        y: 0,
         url: thumbsSprite,
-        height: thumbsHeight,
+        height: this._thumbsHeight,
         width: thumbsWidth
       };
       return new ThumbnailInfo(thumbnailInfo);