+++ src/common/utils/kaltura-params.js
@@ -1,5 +1,6 @@
 //@flow
 import {StreamType, Utils} from '@playkit-js/playkit-js';
+import {getServerUIConf} from './setup-helpers';
 
 const PLAY_MANIFEST = 'playmanifest/';
 const PLAY_SESSION_ID = 'playSessionId=';
@@ -137,12 +138,13 @@ function getQueryStringParamDelimiter(url: string): string {
 
 /**
  * @param {string} url - url
+ * @param {string} productVersion - product version
  * @return {string} - the url with the client tag appended in the query params
  * @private
  */
-function addClientTag(url: string): string {
+function addClientTag(url: string, productVersion: ?string): string {
   if (url.indexOf(CLIENT_TAG) === -1) {
-    url += getQueryStringParamDelimiter(url) + CLIENT_TAG + __VERSION__;
+    url += getQueryStringParamDelimiter(url) + CLIENT_TAG + (productVersion || __VERSION__);
   }
   return url;
 }
@@ -158,6 +160,7 @@ function addKalturaParams(player: Player, playerConfig: PartialKPOptionsObject):
   handleSessionId(player, playerConfig);
   const sources = playerConfig.sources;
   const sessionId = playerConfig.session && playerConfig.session.id;
+  const productVersion = getServerUIConf()?.productVersion;
   Object.values(StreamType).forEach(key => {
     // $FlowFixMe
     if (sources[key]) {
@@ -165,13 +168,13 @@ function addKalturaParams(player: Player, playerConfig: PartialKPOptionsObject):
         if (typeof source.url === 'string' && source.url.toLowerCase().indexOf(PLAY_MANIFEST) > -1 && !source.localSource) {
           source.url = updateSessionIdInUrl(source.url, sessionId);
           source.url = addReferrer(source.url);
-          source.url = addClientTag(source.url);
+          source.url = addClientTag(source.url, productVersion);
         }
         if (source.drmData && source.drmData.length) {
           source.drmData.forEach(drmData => {
             if (typeof drmData.licenseUrl === 'string' && [UDRM_DOMAIN, CUSTOM_DATA, SIGNATURE].every(t => drmData.licenseUrl.includes(t))) {
               drmData.licenseUrl = updateSessionIdInUrl(drmData.licenseUrl, sessionId, DRM_SESSION_ID);
-              drmData.licenseUrl = addClientTag(drmData.licenseUrl);
+              drmData.licenseUrl = addClientTag(drmData.licenseUrl, productVersion);
               drmData.licenseUrl = addReferrer(drmData.licenseUrl);
               drmData.licenseUrl = addUIConfId(drmData.licenseUrl, playerConfig);
             }